+++ native~/Runtime/src/CesiumWgs84EllipsoidImpl.h
@@ -2,14 +2,22 @@
 
 #include <DotNet/Unity/Mathematics/double3.h>
 
+#include <optional>
+
 namespace DotNet::CesiumForUnity {
-class CesiumTransforms;
+class CesiumWgs84Ellipsoid;
 }
 
 namespace CesiumForUnityNative {
 
-class CesiumTransformsImpl {
+class CesiumWgs84EllipsoidImpl {
 public:
+  static DotNet::Unity::Mathematics::double3 GetRadii();
+  static std::optional<DotNet::Unity::Mathematics::double3>
+  ScaleToGeodeticSurface(
+      DotNet::Unity::Mathematics::double3 earthCenteredEarthFixed);
+  static DotNet::Unity::Mathematics::double3 GeodeticSurfaceNormal(
+      DotNet::Unity::Mathematics::double3 earthCenteredEarthFixed);
   static DotNet::Unity::Mathematics::double3
   LongitudeLatitudeHeightToEarthCenteredEarthFixed(
       DotNet::Unity::Mathematics::double3 longitudeLatitudeHeight);