+++ CHANGES.md
@@ -1,5 +1,17 @@
 # Change Log
 
+### v0.2.0
+
+##### Breaking Changes :mega:
+
+- Renamed `CesiumTransforms` to `CesiumWgs84Ellipsoid`.
+
+##### Additions :tada:
+
+- Added `CesiumCameraController`, a globe-aware controller that adapts its speed and clipping planes based on its height from the globe.
+- Added `CesiumFlyToController`, a controller that can smoothly fly to locations across the globe.
+- Added an option to add a `DynamicCamera` from the Cesium panel to the scene. The `DynamicCamera` contains `CesiumCameraController` and `CesiumFlyToController` components and offers easy navigation of the globe.
+
 ### v0.1.2
 
 ##### Fixes :wrench: